Hygiene: Refer to deployment-maps04 in the main 'beta' environment
Current development on this project is targeting Debian Stretch. This
patch updates the 'beta' environment to refer to the new Stretch instance
(deployment-maps04) rather than the legacy Jessie instance (deployment-
maps03). Code can still be deployed to deployemnt-maps03 via a new 'beta-
jessie' environment until the production switchover to Stretch is
complete, at which point it should be dropped and deployment-maps03
destroyed.

Change-Id: I87111c199a30af4282dd83d1536c28ac2a6465a2